+++ changelogger/conf/models.py
@@ -1,11 +1,14 @@
+import tomllib
 from collections import defaultdict
 from pathlib import Path
 from typing import DefaultDict
 
 import yaml  # type: ignore
-from pydantic import BaseModel
+from pydantic import BaseModel, validator
+from semver import VersionInfo
 
-from changelogger.conf._defaults import (
+from changelogger.conf import git
+from changelogger.conf.defaults import (
     CHANGELOGGER_PATH,
     DEFAULT_CHANGELOG_PATH,
     DEFAULT_LINKS_JINJA_PATH,
@@ -13,7 +16,6 @@ from changelogger.conf._defaults import (
     DEFAULT_OVERVIEW_JINJA_PATH,
     DEFAULT_OVERVIEW_JINJA_PATTERN,
 )
-from changelogger.conf.git import get_git_repo
 
 
 class VersionedFile(BaseModel):
@@ -22,6 +24,13 @@ class VersionedFile(BaseModel):
     jinja: str | None = None
     jinja_rel_path: Path | None = None
     context: dict = {}
+
+    def simple_dict(self) -> dict:
+        return {
+            k: v if not isinstance(v, Path) else str(v)
+            for k, v in self.dict().items()
+            if v
+        }
 
 
 class ChangelogSegment(BaseModel):
@@ -53,7 +62,7 @@ class Changelog(BaseModel):
     def as_versioned_files(self) -> list[VersionedFile]:
         context: DefaultDict = defaultdict(dict)
         if self.has_defaults():
-            context["git"]["repo"] = get_git_repo()
+            context["git"] = git.get_ctx()
 
         return [
             VersionedFile(
@@ -66,9 +75,29 @@ class Changelog(BaseModel):
         ]
 
 
+class ProjectMetadata(BaseModel):
+    version: VersionInfo
+    description: str
+
+    class Config:
+        arbitrary_types_allowed = True
+
+    @validator("version", pre=True)
+    def to_version_info(cls, v: str) -> VersionInfo:
+        return VersionInfo.parse(v)
+
+    @classmethod
+    def build(cls) -> "ProjectMetadata":
+        metadata = tomllib.loads(
+            Path("pyproject.toml").read_text(),
+        )
+        return cls(**metadata["tool"]["poetry"])
+
+
 class ChangeloggerConfig(BaseModel):
     changelog: Changelog = Changelog()
     versioned_files: list[VersionedFile] = []
+    metadata: ProjectMetadata = ProjectMetadata.build()
 
     @classmethod
     def from_config_or_default(cls) -> "ChangeloggerConfig":